Use correct timestamp in Note to self envelopes
When sending a SentTranscript sync message manually (as done for note to
self messages), the timestamp of the envelope is set to the current time
instead of the transcript message's timestamp.

This causes these sent message to be rejected when received by
libsignal-service-java:
org.whispersystems.libsignal.InvalidMessageException:
Timestamps don't match: 1553373769631 vs 1553373769692 (ProtocolInvalidMessageException)

It looks like Signal-Desktop doesn't do this validation (which is probably
a bug by itself), so the timestamp mismatch can't be observed when sending
messages from Signal-Android to Signal-Desktop.